backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
langchain-google-genai==0.0.6
langchain==0.1.0
pandas==2.1.4
pydantic==2.5.0
requests==2.31.0